mds: add orchestration to migrate mds daemons to new names
Only mds with names starting with a digit need to be migrated. The first
part of this orchestration deploys the new names (if name do not start
with a digit, nothing is deployed), then the old mds daemons are
stopped. The latter part is guarded with a check so as to not
accidentally stop correct names.

Signed-off-by: Jan Fajerski <jfajerski@suse.com>